requirements.txt
python-twitter
celery
redis
pytz==2013b
pymongo
flask
virtualenvwrapper
